requirements.txt
asgiref==3.8.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.2
dj-rest-auth==7.0.1
Django==5.2
django-allauth==65.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
idna==3.10
psycopg2==2.9.10
pycparser==2.22
PyJWT==2.9.0
requests==2.32.3
sqlparse==0.4.3
tzdata==2025.2
urllib3==2.4.0
